fastapi-projects/16-vidya-rag/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
chromadb==0.5.23
python-multipart==0.0.20
pydantic==2.10.4
